change cookie into token

+++ backend/src/friends/friends.controller.ts
@@ -5,6 +5,7 @@ import {
   Inject,
   Param,
   ParseIntPipe,
+  Req,
 } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { SkipThrottle } from '@nestjs/throttler';
@@ -12,6 +13,7 @@ import { Routes, ServerEvents, Services } from '../utils/constants';
 import { AuthUser } from '../utils/decorators';
 import { User } from '../utils/typeorm';
 import { IFriendsService } from './friends';
+import { AuthenticatedRequest } from 'src/utils/types';
 
 @SkipThrottle()
 @Controller(Routes.FRIENDS)
@@ -23,18 +25,18 @@ export class FriendsController {
   ) {}
 
   @Get()
-  getFriends(@AuthUser() user: User) {
+  getFriends(@Req() req: AuthenticatedRequest) {
     console.log('Fetching Friends');
-    return this.friendsService.getFriends(user.id);
+    return this.friendsService.getFriends(req.userId);
   }
 
   @Delete(':id/delete')
   async deleteFriend(
-    @AuthUser() { id: userId }: User,
+    @Req() req: AuthenticatedRequest,
     @Param('id', ParseIntPipe) id: number,
   ) {
-    const friend = await this.friendsService.deleteFriend({ id, userId });
-    this.event.emit(ServerEvents.FRIEND_REMOVED, { friend, userId });
+    const friend = await this.friendsService.deleteFriend({ id, userId: req.userId });
+    // this.event.emit(ServerEvents.FRIEND_REMOVED, { friend, userId });
     return friend;
   }
 }
